backend/requirements/base.txt
# Core
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3

# Authentication
djangorestframework-simplejwt==5.4.0

# Database
psycopg2-binary==2.9.10
dj-database-url==2.3.0

# Excel Export
openpyxl==3.1.5

# Environment
python-decouple==3.8

# API Documentation
drf-spectacular==0.28.0

# Server
gunicorn==23.0.0
whitenoise==6.8.2
